separate cycle tests
